Added feature to allow backup/restore directories with spaces
To add space in directory use the escape secuence to add space in path (backslash before). This is an example in cya.conf:
BACKUP_DIRECTORIES="/This\ is\ one/a\ test/"
OVERRIDE_BACKUP_DIRECTORIES="/This\ is\ other/second\ test/"
EXCLUDE/This\ is\ other/="exclude\ me/"
BACKUP_FILES="/This\ is\ third/test\ third/cya.test /This\ is\ third/test\ third/cya\ test\ with\ spaces"

Tested save and restore. Works on directories and files with spaces. Also works on EXCLUDE directory option with spaces.